backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langgraph
langchain-core
langchain-openai
langchain-community
websockets
faiss-cpu
networkx
pytest
httpx
pyyaml
playwright
pyautogui
docker
Pillow
sentence-transformers
slowapi
python-jose[cryptography]
keyring
faster-whisper
sounddevice
numpy
openwakeword
piper-tts
tenacity
psutil
python-multipart
easyocr
opencv-python
hypothesis
ruff
mypy
